+++ include/expander.h
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/23 15:41:27 by tsishika          #+#    #+#             */
-/*   Updated: 2023/09/28 22:05:14 by tsishika         ###   ########.fr       */
+/*   Updated: 2023/09/29 01:53:11 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,17 +17,20 @@
 # include "builtin.h"
 # include "lexer.h"
 
+
+// expander_handling.c
+char	*expand_env_string(char	*str, size_t *i, t_env *env_lst);
+
 // epander_utils.c
 bool	is_identifier_char(char c);
 bool	is_expandable(char *str, size_t *i);
 
-// void	str_expander(char **str);
+// expander.c
 void	env_var_expander(t_ast *ast, t_env *env_lst);
 
-char	*no_quote_expander(char *str, size_t *i, t_env *env_lst);
-
-// single_quote_expander.c
+// quote_handling.c
 char	*single_quote_expander(char *str, size_t *i);
-
+char	*no_quote_expander(char *str, size_t *i, t_env *env_lst);
+char	*double_quote_expander(char *str, size_t *i, t_env *env_lst);
 
 #endif
