+++ src/expander/expander.c
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/23 15:42:36 by tsishika          #+#    #+#             */
-/*   Updated: 2023/09/28 22:26:10 by tsishika         ###   ########.fr       */
+/*   Updated: 2023/09/29 01:41:44 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,103 +15,6 @@
 #include "exec.h"
 #include "libft.h"
 #include <libc.h>
-// #include "../../libft/include/libft.h"
-
-// char	*get_env_value(char *str, size_t *i, t_env *env_lst)
-// {
-// 	size_t	j;
-// 	char	*serch_name;
-// 	char	*get_env_value;
-
-// 	j = *i + 1;
-// 	if (str[j] == '$' || str[j] == '\0')
-// 	{
-// 		get_env_value = ft_strdup("$");
-// 		return (get_env_value);
-// 	}
-// 	while (str[j] && str[j] != '$')
-// 		j++;
-// 	serch_name = ft_substr(str, *i + 1, j - (*i + 1));
-// 	if (!serch_name)
-// 		exit(1);
-// 	while (env_lst)
-// 	{
-// 		if (ft_strcmp(serch_name, env_lst->name) == 0)
-// 		{
-// 			get_env_value = ft_strdup(env_lst->value);
-// 			if (!get_env_value)
-// 				exit(0);
-// 			break ;
-// 		}
-// 		env_lst = env_lst->next;
-// 	}
-// 	if (!env_lst)
-// 		get_env_value = ft_strdup("");
-// 	*i = j - 1;
-// 	free(serch_name);
-// 	return (get_env_value);
-// }
-
-char	*get_env_value1(char *str, size_t *i, t_env *env_lst)
-{
-	size_t	j;
-	char	*serch_name;
-	char	*get_env_value;
-
-	j = *i + 1;
-	if (str[j] == '$' || str[j] == '\0')
-	{
-		get_env_value = ft_strdup("$");
-		return (get_env_value);
-	}
-	while (str[j] != DOUBLE_QUOTE && str[j] != '$')
-		j++;
-	serch_name = ft_substr(str, *i + 1, j - (*i + 1));
-	if (!serch_name)
-		exit(1);
-	while (env_lst)
-	{
-		if (ft_strcmp(serch_name, env_lst->name) == 0)
-		{
-			get_env_value = ft_strdup(env_lst->value);
-			if (!get_env_value)
-				exit(0);
-			break ;
-		}
-		env_lst = env_lst->next;
-	}
-	if (!env_lst)
-		get_env_value = ft_strdup("");
-	*i = j - 1;
-	free(serch_name);
-	return (get_env_value);
-}
-
-char	*double_quote_expander(char *str, size_t *i, t_env *env_lst)
-{
-	char	*sub;
-	char	*first;
-	char	*ans;
-	size_t	j;
-
-	j = *i + 1;
-	first = ft_strdup("");
-	while(str[j] != DOUBLE_QUOTE && str[j] != '\0')
-	{
-		if(str[j] == '$')
-			sub = get_env_value1(str, &j, env_lst);
-		else
-			sub = ft_substr(str, j, 1);
-		ans = ft_strjoin(first, sub);
-		free(first);
-		free(sub);
-		first = ans;
-		j++;
-	}
-	*i = j;
-	return (first);
-}
-
 
 void	str_expander(char **str, t_env *env_lst)
 {
