+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/19 17:33:13 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/28 23:29:46 by tsishika         ###   ########.fr       */
+/*   Updated: 2023/09/29 01:38:22 by tsishika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
